Decode colors with the v1 (and questionable) algorithm
the biggest issue with this right now is that it overvalues the opening
rows. (the python code uses `skip_iterator` to circumvent this -- but
since that has other issues I think the lazy approach is best for now)
The second biggest issue is that taking an average of the colors and
comparing the distance at the end is likely a better algorithm. ...
probably.
But for parity purposes with the python code, we'll run with this for
now.