Still a lot of loose ends, but here's a potentially functional encoder
With 6 bits per tile, we have 9300 bytes per frame. With 140:15 ecc (in
155 reed solo chunks), we have 60*140=8400 bytes per frame. With 14 byte
fountain-code metadata per frame, we have 8386 bytes of real data per
frame. With 599 byte fountain (wirehair) blocks -- could also double it
to 1198 (or anything 8386 is divisible by) -- we have 14 data blocks per
frame.

It would be nice to not have to precompute number factorizations for
block sizes. We'll see.