Rolling back a few overambitious (e.g. wrong) optimizations.
Stomping out calls to use opencv's crop was a red herring, so that can
be rolled back... also, the attempt to skip over the first N pixels in
the mean() functions was wrong anyway.
For get_best_symbol's short-circuiting, we'll err on the side of
correctness and stick to the narrowest case for now.