WIP for hack-y undistort.
Basic premise: opencv camera calibration seems to require everything
short of animal sacrifice to get working correctly, and that's not
particularly viable for a cell phone app. The "right answer" is probably
a database of precomputed cameras+distortion matrices, but that's also
out of scope for right now. So: hax

Have some off-by-one errors (or discrepancies) in the Scanner which
compound through the edge detection and into the calibration. We'll see
how bad that is.