add zstd padding to push small files over the limit for wirehair
It's not clear to me what to do with the small files case. Ideally there
would be some additional error correction with the otherwise-unused
bytes -- and in some sense, there is (in that we have 10 fountain
packets in one image). But there's also a ton of dead space -- the
majority of the image, in fact.

An option -- maybe the *right* option -- is to have a separate
encoding/decoding scheme for small data. But that is complexity I'm not
interested in right now.