Trying the wasm experiment
I have this (kind of) working on a very-tortured libcimbar repo -- this
is a first pass at trying to integrate it into the project hierarchy.

Ideally we could have CI builds both run normally, and in WASM mode, to
catch weirdness. There are currently no suitable tests for the
asm.js/WASM stuff, however...