Set CellDrift limit to 7 -- improves decoding of distorted images.
We're basically trusting the flood decode to do a good job to keep us on
the rails, but it turns out that it *does* do a pretty good job.

Giving the decoder more leash to handle drift also means the "undistort"
step is less important/relevant, which is good because (1) it's not very
effective and (2) it's not very fast.