Not sure about this rotate api, but we need a way to reset rotation...
Anyway, like the shakycam trick, the idea here is to have successive
cimbar frames misalign their anchor pattern -- in this case by rotating
the image 90 degrees each frame. The misalignment allows the scanner
(receiver) to discard partial or ghosted frames without the expense of a
full decode.

The rotate seems to perform *somewhat* similarly to the shakycam, but
not confident about that yet.

One obvious advantage is that rotation doesn't require changing the
canvas size -- so this seems like a good fit for cimbar.js.