This seems like something that ought to work.
If you change ecc settings, or the overall storage capacity of the image
(e.g. choosing an 8 bit encode instead of a 4 bit), we need a new
FountainEncoder object... But we don't necessarily need a new
fountain_encoder_stream()!

There is still a gotcha -- because of the way small files are currently
handled (editorial note: "poorly") -- if the file is smaller than a
fountain chunk, and we go to a higher capacity encoding (e.g. from 6
bits per tile to 7), the encoder will choke and you'll have to re-submit
the file. Not sure there's a good way around this right now...