This seems ok?
Not 100% confident about this api, but the idea is that we create a new
fountain_encoder_stream from an old one -- using a new `buffer_size`. We
std::move() the underlying data. Also, in the case where the buffer_size
is larger than the underlying data (this can happen due to our reliance
on padding for small files), we bail out.