configurable redundancy for encode_fountain().
It's not obvious what the right multiplier actually is, but it will vary
by use case.
1. temporary outputs to video. Benefits from a high total number of
frames. If the decode rate is a % of the overall frame number, decoding
speed will degrade with the number of duplicate frames we decode. So the
optimal number is to always have a frame we've never seen before, e.g.
an infinite rendundancy factor. I settled on 4, but it could be higher.
2. outputs to file. Given finite resources -- e.g. disk space, or paper
pages -- decoder flexibility is not as much of a priority as having the
full data and some level of extra frames if some are lost. The number
here is 1.2, but anything over 1 is probably fine?