Experimenting with ways to improve the flood decode
The idea here is that there are some circumstances we can do better than
"we think the 4 adjacent tiles are here".
If we see a line of tiles that are right on the money drift wise
(i.e. they require no adjustment), we can be reasonably confident we're
doing ok, and take the step of looking ahead in either direction (in
practice one of these will be a no-op, since we *came* from that direction)

This seems to lead to a small-but-consistent improvement in decode quality.