Experiment with a different color calculation?
Instead of averaging everything, or cutoffs (I experimented with
cutoffs), this is using the nlargest for each color channel. Could also
use N brightest pixels, but this is simpler to calculate. (probably
worth investigating if it's better though...)

N=9 pixels right now. We're sampling 18px for the 8x8 (only sampling 9px
for the 5x5), so we're filtering out the dimmest 50%. (based on some
trial+error, N=5 seems ok for the 5x5 case)