WIP to vary fountain chunk size by number of symbol bits...
Reduces the number of magical config numbers (at the cost of more
complexity in passing around bits_per_symbol everywhere).

I'd like to do the same with interleave_partitions, *I think*.

When all of this is working, we should be able to cleanly decouple the
symbol+color sections of each cimbar frame.