Change params for symbol decode's adaptiveThreshold operation
The negative constant (-5) with blockSize=3 is an overfit for 8x8 cimbar
based on my current (371 frame) trial set.

blockSize=3 or 5 or 7 aren't *too* different if the subtraction constant
is set to 0 (and intuitively maybe it *should* be).
7 seems slighly worse than the other two, and 5 (marginally) the best.

When the sharpen kernel is enabled, 7 and 5 are more or less identical
(3 is worse), but 7 seems best