Add FountainHeader logic to CimbReader
The premise here is to use the symbol decode to compute the expected
fountain headers for the upcoming color decode -- and then use those
known values to compute the average real-world color values for each
relevant tile in the image.

We can then use those averages to compute a more accurate color
correction matrix.