For encodes via the cimbar cli, skip over frames that will fail extract
Some cimbar frames are bad and will fail extraction due to the data
aligning itself to match the anchor pattern.

This is a design bug/quirk that doesn't matter *too* much for the
on-the-fly "encode data on one screen, decode via camera" use case.
We're constantly generating new frames, after all. But for any offline,
more "QR-code" like usage -- for example printing -- having a bad image
is much more troublesome.

The fix/workaround is to check the generated image after we generate,
and make sure it extracts. If it doesn't, we skip that frame and again.

This is currently only used the cli (and various test code), due to
potential performance and infinite-loop concerns.