At long last, change the default to mode B + some gnarly rotation logic
We'd like to support non-square cimbar configs (more screen space used =
more available bandwidth, probably), and this allows the encoder to do
so -- whiler also working in portrait mode.

There is a truly kludgy and hideous javascript+css hack here to get the
transform right -- for reasons that elude me, rotating around the center
doesn't actually do so. This might be related to display:block or some
other inane css interaction.

There's a less hideous but still annoying js+css hack to get the helper
text ("click here to start") to stay somewhat centered when it rotates.
It's probably good enough for now.

Note: the rotation logic will *not* be in use with the existing mode
B/4C.