add some plumbing to store a string (filename) in a zstd skippable frame
This will go at the start of the compressed file, and we'll check for it
at the start of decompression.

If anything fails (filename is malicious, wrong magic number at the
start, no header at all) we'll fall back to the existing behavior --
which is important, since any old encoder won't encode the header.