When using compression, store the filename on encode, retrieve on decode
Not fully wired up to the user yet (cimbar_recv.js doesn't use it yet,
the decoder sink doesn't actually expose it...), but this is 98% of the
way there.

The filename is itself not compressed, which seems silly and will
require another look.

Includes:
* tweaks to roundtrip tests. This includes exposing an interesting/weird
bug I'd heard hints of but had never reproduced...
* a separate function for the fountain_decoder_sink which is zstd aware.
Since we need to check for the filename, we need to either do it in
advance of opening the stream (this approach), or do a lazy open as part
of the first write() call. For now I went with parsing the name in
advance, since this is essentially an informal extension to the zstd
format. ("when the 0th frame of the file is a skippable frame, and it's
using the first magicVal, and the first byte of the payload is 1, then
the remaining bytes should be treated as the filename") The decompressor
remains "stupid" as a result, e.g. it technically should work for
decompressing any/all zstd data.
* cimbar_send now encodes the src filename
* added const a few places (unrelated to the rest of this, but annoyed
me)
* cimbard_get_filename() added to the decode c api. This could also go
in cimbarz_*, and it's where I put it initially. But the reasoning is
the same as above, e.g. it's not technically part of zstd but an
extension (sort of "cimbar flavored zstd")