have fountain decoder sink surface its filenames + "cimbar_recv2"
For better or worse, we have two different styles of decoder: the CFC
model, where the fds writes the files to disk (in a temp loc on android,
which is then copied on demand), and the webapp model where the decoded
file is ephemeral and passed to the caller in chunks.

Unit tests only take us so far, so cimbar_recv exists to provide a
facsimile of the production use case. So: maybe we need two cimbar_recvs

Anyway, this is the pre-c-api-ified version from commit
8ee997d4646bc5bba2f3673bdfc7527a97574d68. The c-api version has been
relocated to "cimbar_recv2".