WIP: rewrite of config checks to use static threadlocal
We know we need to set defaults somewhere, so that's table stakes.

The options for overrides are:
* pass them around piecemeal, with each function/class that cares about
a variable accepting an override for it. We want *some* of this for
testing purposes, but it's become a nuisance.
* pass a big config object around everywhere. Disgusting
* set the config in a global or threadlocal, and have the various places
treat it as the single source of truth.

The threadlocal vs global choice has its own tradeoffs, but I like not
having to worry about concurrency issues.