Lazy eval the config inside the Decoder::decode() functions
...rather than caching all the params at Decoder() init.

We use Decoder as a long-lived object (aside: not currently reflected in
cimbar_recv_js, which throws it away and creates a new one each call),
which means any variables it stores at the class level should be either
immutable or configurable. For now, we'll say the former -- and pull in
the config only when we need it.

In practice for the android logic, we'll be calling decode() on multiple
threads against the same Decoder, and each of these calls will defer to
the threadlocal Config object to tell it how to do its business.