Reorganizing the encoder c api to (hopefully) lessen OOM errors
The gist is that we currently have to read the whole file into
contiguous memory before handing it off the API. *Additionally*, as a
hard requirement for wirehair, the *compressed* file needs to exist in
contiguous memory.

You might think (as I did) that ~15 MB would not be enough to run afoul
of the browser memory constraints, but apparently this is not so.

We can't do anything about the second (post compressed) requirement, but
the first can be relaxed if the API accepts files in chunks. So:

* cimbare_init_encode() <-- to initialize the encode. Provide filename,
set encode_id (or -1 to auto-increment)
* cimbare_encode_bufsize() <-- the desired chunk size. Currently a
constant.
* cimbare_encode() <-- to provide data. When the data size != the chunk
size, the encode is ready. (this also means we can still achieve the old
behavior of providing the whole file at once)

cimbar_encode(null, 0) can act as a flush() to make sure the encode is
ready -- though this is only relevant when the filesize is a multiple of
the chunk size.