Add back the encoder color hack, and make it toggleable at api level
These have been tested for 4-color mode, so will only be used there.

There are two tricks here:
1. slightly brighten the tone used for magenta, to more closely match
the brightness of the other colors. Perceived brightness of a backlit
display is heavily influenced by green, and magenta (color 3) is the
only one without any green in it. This can result in worse symbol
decodes (as the magenta colored tiles will be darker and sometimes less
visible).
2. add a slight tint to the background color. The idea here is to
slightly color the background, but not too much-- it still needs to
maintain a strong contrast to not interfere with symbol decodes. But a
hint of e.g. red behind the yellow "moves" it away from green and cyan,
resulting in less errors.

After some testing, I'm hoping to enable these tricks by default. But
there will probably be a toggle to turn them off just in case they cause
trouble...