Enable mode Bw in the web ui
This is the fastest cimbar config yet. (ignoring the eternally WIP 5x5
configs...)

But I'm not 100% sold yet on keeping this one around long term. Unlike
mode B and Bm, this basically exists as a "go fast" option, and mode S
*will* be faster if I ever get it working nicely enough.

Still, it feels silly to have a faster mode figured out but
hidden/unusable. So for the time being: mode Bw.