5x5 2 color config
This might be a dead end, but *seems* to function ok. Going down to 2
colors from 4 helps reliability a lot, but the data density gets a lot
closer to 8x8 4c -- with extra drawbacks. As expected, decoding CPU
usage is pretty brutal with all these cache misses.

Going to try a smaller 5x5 config to see if it has any advantages over
mode Bm...